refactor: extract useSlashMenu hook from App.tsx
